Add files via upload
This notebook implements the time series decomposition of the SWAN-SF data using the Discrete Fourier Transform, Haar Wavelet, Symlet and Daubechies Wavelet to train the TimeSeriesForest model for binary prediction of solar flares . It also converts the data to non-time series data to compare performance with time series data across different dimensions of the data. The non-time series data is trained using a RandomForest Ensemble.